Updated theme data freshness
